+++ src/Telegram.Bot/TelegramBotClient.cs
@@ -34,8 +34,6 @@ namespace Telegram.Bot
 
         private readonly string _token;
 
-        private bool _invalidToken;
-
         private readonly HttpClient _httpClient;
 
         #region Config Properties
@@ -157,9 +155,6 @@ namespace Telegram.Bot
         /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invalid</exception>
         public TelegramBotClient(string token, HttpClient httpClient = null)
         {
-            if (!Regex.IsMatch(token, @"^\d*:[\w\d-_]{35}$"))
-                throw new ArgumentException("Invalid token format", nameof(token));
-
             _token = token;
             _baseRequestUrl = $"{BaseUrl}{_token}/";
             _httpClient = httpClient ?? new HttpClient();
@@ -173,9 +168,6 @@ namespace Telegram.Bot
         /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invalid</exception>
         public TelegramBotClient(string token, IWebProxy webProxy)
         {
-            if (!Regex.IsMatch(token, @"^\d*:[\w\d-_]{35}$"))
-                throw new ArgumentException("Invalid token format", nameof(token));
-
             var httpClientHander = new HttpClientHandler
             {
                 Proxy = webProxy,
@@ -194,9 +186,6 @@ namespace Telegram.Bot
             IRequest<TResponse> request,
             CancellationToken cancellationToken = default)
         {
-            if (_invalidToken)
-                throw new ApiRequestException("Invalid token", 401);
-
             string url = _baseRequestUrl + request.MethodName;
 
             var httpRequest = new HttpRequestMessage(request.Method, url)
@@ -241,8 +230,6 @@ namespace Telegram.Bot
                 case HttpStatusCode.OK:
                     break;
                 case HttpStatusCode.Unauthorized:
-                    _invalidToken = true;
-                    throw new ApiRequestException("Invalid token", 401);
                 case HttpStatusCode.BadRequest when !string.IsNullOrWhiteSpace(responseJson):
                 case HttpStatusCode.Forbidden when !string.IsNullOrWhiteSpace(responseJson):
                 case HttpStatusCode.Conflict when !string.IsNullOrWhiteSpace(responseJson):
@@ -278,7 +265,8 @@ namespace Telegram.Bot
                 await GetMeAsync(cancellationToken).ConfigureAwait(false);
                 return true;
             }
-            catch (HttpRequestException e) when (e.Message.EndsWith("404 (Not Found)."))
+            catch (ApiRequestException e)
+                when (e.ErrorCode == 401)
             {
                 return false;
             }
@@ -293,9 +281,6 @@ namespace Telegram.Bot
         public void StartReceiving(UpdateType[] allowedUpdates = null,
             CancellationToken cancellationToken = default)
         {
-            if (_invalidToken)
-                throw new ApiRequestException("Invalid token", 401);
-
             _receivingCancellationTokenSource = new CancellationTokenSource();
 
             cancellationToken.Register(() => _receivingCancellationTokenSource.Cancel());
@@ -312,20 +297,16 @@ namespace Telegram.Bot
             while (!cancellationToken.IsCancellationRequested)
             {
                 var timeout = Convert.ToInt32(Timeout.TotalSeconds);
+                Update[] updates = default;
 
                 try
                 {
-                    var updates =
-                        await
-                            GetUpdatesAsync(MessageOffset, timeout: timeout, allowedUpdates: allowedUpdates,
-                                    cancellationToken: cancellationToken)
-                                .ConfigureAwait(false);
-
-                    foreach (var update in updates)
-                    {
-                        MessageOffset = update.Id + 1;
-                        OnUpdateReceived(new UpdateEventArgs(update));
-                    }
+                    updates = await GetUpdatesAsync(
+                       MessageOffset,
+                       timeout: timeout,
+                       allowedUpdates: allowedUpdates,
+                       cancellationToken: cancellationToken
+                    ).ConfigureAwait(false);
                 }
                 catch (OperationCanceledException)
                 {
@@ -337,6 +318,12 @@ namespace Telegram.Bot
                 catch (Exception generalException)
                 {
                     OnReceiveGeneralError?.Invoke(this, generalException);
+                }
+
+                foreach (var update in updates)
+                {
+                    OnUpdateReceived(new UpdateEventArgs(update));
+                    MessageOffset = update.Id + 1;
                 }
             }
 
@@ -928,7 +915,8 @@ namespace Telegram.Bot
             MakeRequestAsync(new EditInlineMessageTextRequest(inlineMessageId, text)
             {
                 DisableWebPagePreview = disableWebPagePreview,
-                ReplyMarkup = replyMarkup
+                ReplyMarkup = replyMarkup,
+                ParseMode = parseMode
             }, cancellationToken);
 
         /// <inheritdoc />